rrr: convert rpc/idempotency to module (rrr.idempotency)
idempotency.cpp module unit; .hpp becomes one-line shim. Exports
IdempotencyKey POD + IdempotencyKeyHash + Marshal stream operators +
IdempotencyConfig with 4 presets + CachedResponse + IdempotencyKeyGenerator
+ IdempotencyCache (LRU + TTL + stats).

Imports rrr.marshal (Marshal class for cache value bytes),
rrr.serializable, rrr.threading.

31 modules total.

Measurement (rrr + rpcbench, -j32, clean):
  +any_message:    56.85 s, ~351 MB PCM
  +idempotency:    67.62 s, ~XXX MB PCM

+10.8 s wallclock — also wide fan-out: idempotency is imported via
client.hpp which most TUs include. 4.01× baseline; under 5× ceiling.
